pyproject.toml
[project]
name = "wqk-mcp"
version = "0.1.0"
description = "Stock Analysis MCP Server - A comprehensive stock market data analysis tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core MCP framework
    "fastmcp>=2.9.2",
    "mcp>=1.9.1",
    
    # Data processing and analysis
    "pandas>=2.2.3",
    "numpy>=2.2.6",
    "scikit-learn>=1.6.1",
    
    # Financial data sources
    "akshare>=1.16.95",
    "adata>=2.9.2",
    
    # HTTP clients and web scraping
    "httpx>=0.28.1",
    "requests>=2.32.3",
    "crawl4ai>=0.6.3",
    "playwright>=1.52.0",
    
    # Search and AI tools
    "tavily-python>=0.7.2",
    
    # Configuration and utilities
    "python-dotenv>=1.1.0",
    "pydantic>=2.11.5",
    "typing-extensions>=4.13.2",
    
    # File processing
    "openpyxl>=3.1.5",
    "xlrd>=2.0.1",
    
    # Async support
      "aiofiles>=24.1.0",
  ]

  [project.optional-dependencies]
  dev = [
      "pytest>=7.0.0",
      "pytest-asyncio>=0.21.0",
      "black>=23.0.0",
      "isort>=5.12.0",
      "flake8>=6.0.0",
      "mypy>=1.0.0",
  ]

  [project.scripts]
  wqk-mcp = "server:main"

  [build-system]
  requires = ["hatchling"]
  build-backend = "hatchling.build"

  [tool.hatch.build.targets.wheel]
  packages = ["."]

 [tool.uv]
 dev-dependencies = [
     "pytest>=7.0.0",
     "pytest-asyncio>=0.21.0",
     "black>=23.0.0",
     "isort>=5.12.0",
     "flake8>=6.0.0",
     "mypy>=1.0.0",
 ]

 [tool.black]
 line-length = 88
 target-version = ['py313']

 [tool.isort]
 profile = "black"
 line_length = 88

 [tool.mypy]
 python_version = "3.13"
 warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = true
